SemanticModel: Use binding position in resolve_load
This means that the binding resolved on a deferred pass is the actual binding
made at that point in the scope, not the eventual binding at the end of the
scope.

This means that names bound and later deleted can still be resolved from a
deferred pass as defined, even if they are later deleted.